refactor main library modules, transfer to new crate
