--- jcli/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod jcli_lib;
-pub use crate::jcli_lib::*;
